Fixed Memory Leak
Fixed memory leak that turned out to be with the = operator in the ObjectHandler.  It created a new array but failed to delete the old one, which I thought was not a problem since the default only had a float of 1.  However, it overwrites the old ones as well, which have size 1200.