Code Cleanup and Main Test File
Cleaned up the code a little bit and recommitted the main.cpp file so that testing can be run.  The main outputs to a file the colors of the pixels with each pixel on a new line so that it can be read.  It currently renders two triangles with u,v colors.