Added UV Calculations
The program now calculates the texture coordinates for each triangle by calculating the UV values and adjusting them based off of the individual UV texture coordinates of each vertice.  This is in preparation for having a full blown texture system.  Current output of the program is a triple (u, v, 0) where u and v are the calculated texture coordinates.