Updated Device To Use Hits
The rendering process now uses hits instead of raw values.  This makes it far more intuitive.  Additionally, some changes were made to the the Hit struct to make the default values of hit and object to be more logical.  In its new state it will fix the issue of 'hit' defaulting to true and causing garbage values to be interpreted as real.